feat(version): v0.1.0

+++ app/Http/Controllers/UserController.php
@@ -82,6 +82,9 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if ($user->role == 'admin') {
+            return redirect()->back()->withErrors('ადმინის წაშლა შეუძლებელია');
+        }
         $user->delete();
 
         return redirect()->back();
